requirements.txt
flask
pywebview
pyinstaller
